feat: Selesai Fase 5 - Menyelesaikan fitur checkout, riwayat pesanan, dan pengurangan stok otomatis

+++ database/migrations/2026_07_20_135614_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         // Menghubungkan produk ke kategori. Jika kategori dihapus, produk di dalamnya otomatis ikut terhapus
         $table->foreignId('category_id')->constrained()->onDelete('cascade'); 
         $table->string('nama'); // Nama Laptop/Aksesoris
-        $table->integer('harga');
+        $table->bigInteger('harga');
         $table->integer('stok');
         $table->string('foto')->nullable(); // Tempat menyimpan path file foto produk
         $table->text('deskripsi')->nullable();
